+++ blueprint/billing-base/server.ts
@@ -9,6 +9,7 @@ import { checkoutSessionRouter } from './api/routes/checkoutSession.routes';
 import { paymentLinkRouter } from './api/routes/paymentLink.routes';
 import { planRouter } from './api/routes/plan.routes';
 import { subscriptionRouter } from './api/routes/subscription.routes';
+import { complianceRouter } from './api/routes/compliance.routes';
 import { ci, tokens } from './bootstrapper';
 import { billingSdkClient } from './sdk';
 
@@ -43,6 +44,7 @@ app.use(checkoutSessionRouter);
 app.use(paymentLinkRouter);
 app.use(planRouter);
 app.use(subscriptionRouter);
+app.use(complianceRouter);
 
 //! registers the sdk client
 app.registerSdks(billingSdkClient);