+++ cli/src/core/package_json/package_json_constants.rs
@@ -232,13 +232,13 @@ pub(crate) const APP_MONITORING_VERSION: &str = "workspace:*";
 // @forklaunch/bunrun
 pub(crate) const BUNRUN_VERSION: &str = "~0.0.2";
 // @forklaunch/common
-pub(crate) const COMMON_VERSION: &str = "~1.1.0";
+pub(crate) const COMMON_VERSION: &str = "~1.1.1";
 // @forklaunch/core
-pub(crate) const CORE_VERSION: &str = "~1.1.0";
+pub(crate) const CORE_VERSION: &str = "~1.1.1";
 // @forklaunch/express
-pub(crate) const EXPRESS_VERSION: &str = "~1.1.0";
+pub(crate) const EXPRESS_VERSION: &str = "~1.1.1";
 // @forklaunch/hyper-express
-pub(crate) const HYPER_EXPRESS_VERSION: &str = "~1.1.0";
+pub(crate) const HYPER_EXPRESS_VERSION: &str = "~1.1.1";
 // @forklaunch/implementation-billing-base
 pub(crate) const BILLING_BASE_VERSION: &str = "~1.0.1";
 // @forklaunch/implementation-billing-stripe
@@ -254,9 +254,9 @@ pub(crate) const WORKER_DATABASE_VERSION: &str = "~1.0.1";
 // @forklaunch/implementation-worker-kafka
 pub(crate) const WORKER_KAFKA_VERSION: &str = "~1.0.1";
 // @forklaunch/infrastructure-redis
-pub(crate) const INFRASTRUCTURE_REDIS_VERSION: &str = "~1.1.0";
+pub(crate) const INFRASTRUCTURE_REDIS_VERSION: &str = "~1.1.1";
 // @forklaunch/infrastructure-s3
-pub(crate) const INFRASTRUCTURE_S3_VERSION: &str = "~1.1.0";
+pub(crate) const INFRASTRUCTURE_S3_VERSION: &str = "~1.1.1";
 // @forklaunch/interfaces-billing
 pub(crate) const BILLING_INTERFACES_VERSION: &str = "~1.0.1";
 // @forklaunch/interfaces-iam
@@ -264,13 +264,13 @@ pub(crate) const IAM_INTERFACES_VERSION: &str = "~1.0.1";
 // @forklaunch/interfaces-worker
 pub(crate) const WORKER_INTERFACES_VERSION: &str = "~1.0.1";
 // @forklaunch/internal
-pub(crate) const INTERNAL_VERSION: &str = "~1.1.0";
+pub(crate) const INTERNAL_VERSION: &str = "~1.1.1";
 // @forklaunch/testing
-pub(crate) const TESTING_VERSION: &str = "~1.1.0";
+pub(crate) const TESTING_VERSION: &str = "~1.1.1";
 // @forklaunch/universal-sdk
-pub(crate) const UNIVERSAL_SDK_VERSION: &str = "~1.1.0";
+pub(crate) const UNIVERSAL_SDK_VERSION: &str = "~1.1.1";
 // @forklaunch/validator
-pub(crate) const VALIDATOR_VERSION: &str = "~1.1.0";
+pub(crate) const VALIDATOR_VERSION: &str = "~1.1.1";
 // @mikro-orm/core
 pub(crate) const MIKRO_ORM_CORE_VERSION: &str = "7.0.5";
 // @mikro-orm/migrations