+++ framework/hyper-express/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @forklaunch/hyper-express
 
+## 1.1.1
+
+### Patch Changes
+
+- add compliance utilities
+- Updated dependencies
+  - @forklaunch/validator@1.1.1
+  - @forklaunch/common@1.1.1
+  - @forklaunch/core@1.1.1
+  - @forklaunch/ws@1.1.1
+
 ## 1.1.0
 
 ### Minor Changes