+++ framework/infrastructure/redis/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @forklaunch/infrastructure-redis
 
+## 1.1.1
+
+### Patch Changes
+
+- add compliance utilities
+- Updated dependencies
+  - @forklaunch/common@1.1.1
+  - @forklaunch/core@1.1.1
+
 ## 1.1.0
 
 ### Minor Changes